# docs/swagger.json
{
    "openapi": "3.1.0",
    "info": {
        "version": "1.0.0",
        "title": "Conatcs Node.js",
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "description": "Conatcs app"
    },
    "servers": [
        {
            "url": "http://localhost:3000"
        },
        {
            "url": "https://nodejs-hw-mongodb-rghb.onrender.com"
        }
    ],
...
    "paths": {
...
        "/contacts/{Id}": {
            "get": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Get contact by id",
                "operationId": "getContact",
                "description": "Get contact by id",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "65ca67e7ae7f10c88b598384"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get contact by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully found contact with id 65ca67e7ae7f10c88b598384!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/contact"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Update contact",
                "operationId": "updateContacts",
                "description": "Update a contact",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "65ca67e7ae7f10c88b598384"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "description": "Contacts name",
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phoneNumber": {
                                        "description": "Contacts phone number",
                                        "type": "string",
                                        "example": "1234567890"
                                    },
                                    "email": {
                                        "description": "Contacts email",
                                        "type": "string",
                                        "example": "example@gmail.com"
                                    },
                                    "isFavourite": {
                                        "description": "Contacts favourite status",
                                        "type": "boolean",
                                        "example": false
                                    },
                                    "contactType": {
                                        "description": "Type of contact",
                                        "type": "string",
                                        "enum": [
                                            "work",
                                            "home",
                                            "personal"
                                        ]
                                    },
                                    "photo": {
                                        "description": "Conatcs photo",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updates a contact",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully updated a contact!"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/contact"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Contacts"
                ],
                "summary": "Delete contact by id",
                "operationId": "deleteContact",
                "description": "Delete contact",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "Id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "65ca67e7ae7f10c88b598384"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Delete contact by id",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "required": [
                                        "status",
                                        "message",
                                        "data"
                                    ],
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 204
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully deleted contact with id 65ca67e7ae7f10c88b598384"
                                        },
                                        "data": {
                                            "type": "object",
                                            "$ref": "#/components/schemas/contact"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "contact": {
                "type": "object",
                "required": [
                    "_id",
                    "name",
                    "phoneNumber",
                    "contactType",
                    "userId"
                ],
                "properties": {
                    "_id": {
                        "description": "contacts id",
                        "type": "string",
                        "example": "675c8399c0dd8a4ccbe173fd"
                    },
                    "name": {
                        "description": "contacts name",
                        "type": "string",
                        "example": "John Doe"
                    },
                    "phoneNumber": {
                        "description": "contacts phoneNumber",
                        "type": "number",
                        "example": 380672222222
                    },
                    "isFavourite": {
                        "description": "contacts favourite status",
                        "type": "boolean",
                        "example": true
                    },
                    "email": {
                        "description": "Contact email address",
                        "type": "string",
                        "example": "john.doe@example.com"
                    },
                    "contactType": {
                        "description": "contacts type",
                        "type": "string",
                        "enum": [
                            "work",
                            "home",
                            "personal"
                        ]
                    },
                    "photo": {
                        "description": "Link to contacts photo",
                        "type": "string",
                        "example": "https://res.cloudinary.com/uqwgdu/image/upload/image.png"
                    },
                    "userId": {
                        "description": "ID of the user who owns this contact",
                        "type": "string",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    }
                }
            }
        },
        "responses": {
            "401": {
                "description": "Unauthorized Error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "required": [
                                "status",
                                "message",
                                "data"
                            ],
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 401
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Unauthorized Error"
                                },
                                "data": {
                                    "type": "object",
                                    "required": [
                                        "message"
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Access token not provided"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "404": {
                "description": "Not found error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "required": [
                                "status",
                                "message",
                                "data"
                            ],
                            "properties": {
                                "status": {
                                    "type": "integer",
                                    "example": 404
                                },
                                "message": {
                                    "type": "string",
                                    "example": "NotFoundError"
                                },
                                "data": {
                                    "type": "object",
                                    "required": [
                                        "message"
                                    ],
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Contact not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}